+++ domaps/domaps.py
@@ -32,102 +32,7 @@ import inspect
 import warnings
 from itertools import cycle, combinations
 
-
-class ConstantsClass(type):
-    """A metaclass for classes that should only contain string constants."""
-
-    def __setattr__(self, name, value):
-        raise TypeError("Constants class cannot be modified")
-
-    def get_values(cls):
-        """Get all user-defined string values of the class."""
-        return [
-            value
-            for key, value in cls.__dict__.items()
-            if not key.startswith("__") and isinstance(value, str)
-        ]
-
-
-class DataFrameStrings(metaclass=ConstantsClass):
-    # Indexing levels
-    EXPERIMENT = "Experiment"  # Refers to a single set of maps that can be loaded into a SpatialDataSet
-    MAP = "Map"  # Refers to a single set of samples that make up one map
-    FRACTION = "Fraction"  # Refers to a fraction of the experimental gradient
-    SET = "Set"  # Refers to a set of column that contain the same kind of data
-    EXP_MAP = "Exp_Map"  # A combination of experiment and map used only in the SpatialDataSetComparison class
-
-    # Column names
-    PROTEIN_IDS = "Protein IDs"
-    GENE_NAMES = "Gene names"
-    ORIGINAL_PROTEIN_IDS = "Original Protein IDs"
-    COMPARTMENT = "Compartment"
-    CLUSTER = "Cluster"
-
-    # Set names
-    LFQ_INTENSITY = "LFQ intensity"
-    MSMS_COUNT = "MS/MS count"
-    INTENSITY = "Intensity"
-    ABUNDANCE = (
-        "Abundance"  # This is the generalizing one, that is used within preprocessing
-    )
-    RATIO = "Ratio"
-    RATIO_COUNT = "Ratio count"
-    RATIO_VARIABILITY = "Ratio variability"
-    NORMALIZED_PROFILE = "normalized profile"
-    LOG_PROFILE = "log profile"
-
-
-class SettingStrings(metaclass=ConstantsClass):
-    # These are various strings used within settings files to define the processing pipeline. Changes make ols settings files incompatible.
-
-    # Filter names
-    FILTER_SILAC_COUNTVAR = "SILAC_countvar"
-    FILTER_MSMS_COUNT = "msms_count"
-    FILTER_CONSECUTIVE = "consecutive"
-
-    # Normalization options
-    NORMALIZATION_SUM = "sum"
-    NORMALIZATION_MEDIAN = "median"
-
-    # Acquisition modes
-    SILAC_ACQUISITION = "SILAC"
-    LFQ_ACQUISITION = "LFQ"
-    INTENSITY_ACQUISITION = "Intensity"
-    CUSTOM_ACQUISIITION = "custom"
-
-    # Preconfigured source modes
-    MAXQUANT_PROTEINS_PIVOT = "MaxQuant_proteins_pivot"
-    SPECTRONAUT_PROTEINS_LONG = "Spectronaut_proteins_long"
-
-    # Acquisition mode settings used to define which sets should be read and how they should be processed
-    INPUT_INVERT = "input_invert"
-    INPUT_LOGGED = "input_logged"
-    INPUT_SAMPLENORMALIZATION = "input_samplenormalization"
-    QUALITY_FILTER = "quality_filter"
-    AVERAGE_MSMS_COUNTS = "average_MSMS_counts"
-    CONSECUTIVE = "consecutive"
-    RATIOCOUNT = "RatioCount"
-    RATIOVARIABILITY = "RatioVariability"
-
-    # Source settings used to define how the data should be read and formatted
-    ORIGINAL_PROTEIN_IDS = "original_protein_ids"
-    GENES = "genes"
-    COLUMN_FILTERS = "column_filters"
-    ANNOTATION_COLUMNS = "columns_annotation"
-    SAMPLES = "samples"
-    ORIENTATION = "orientation"
-
-    # General settings
-    SETS = "sets"
-    SOURCE = "source"
-    ACQUISITION = "acquisition"
-    COMMENT = "comment"
-    NAME_PATTERN = "name_pattern"
-    ORGANISM = "organism"
-    REANNOTATE = "reannotate"
-    REANNOTATION_SOURCE = "reannotation_source"
-    ORGANELLES = "organelles"
-    COMPLEXES = "complexes"
+from domaps.constants import DataFrameStrings, SettingStrings
 
 
 def natsort_index_keys(x):
@@ -447,16 +352,16 @@ class SpatialDataSet:
                     settings[SettingStrings.REANNOTATE],
                     settings[SettingStrings.REANNOTATION_SOURCE],
                 ),
-                organism=settings["organelles"],
+                organism=settings[SettingStrings.ORGANELLES],
                 **{
                     k: v
                     for k, v in settings.items()
                     if k
                     not in [
-                        "acquisition",
+                        SettingStrings.ACQUISITION,
                         SettingStrings.REANNOTATION_SOURCE,
                         SettingStrings.REANNOTATE,
-                        "organism",
+                        SettingStrings.ORGANISM,
                     ]
                 },
             )
@@ -799,15 +704,17 @@ class SpatialDataSet:
 
         # poopulate analysis summary dictionary with (meta)data
         self.analysis_summary_dict["0/1 normalized data"] = df_01_comparison.to_json()
-        self.analysis_summary_dict["organelles"] = self.df_organellarMarkerSet.to_json()
-        self.analysis_summary_dict["complexes"] = self.markerproteins
+        self.analysis_summary_dict[SettingStrings.ORGANELLES] = (
+            self.df_organellarMarkerSet.to_json()
+        )
+        self.analysis_summary_dict[SettingStrings.COMPLEXES] = self.markerproteins
         # self.analysis_summary_dict["changes in shape after filtering"] = shape_dict.copy()
         # add more settings here
         analysis_parameters = {
-            "acquisition": self.acquisition,
+            SettingStrings.ACQUISITION: self.acquisition,
             "filename": self.filename,
-            "comment": self.comment,
-            "organism": self.organism,
+            SettingStrings.COMMENT: self.comment,
+            SettingStrings.ORGANISM: self.organism,
             "processing steps": "\n".join(processing_steps),
             "legacy": self.legacy,
         }
@@ -2644,11 +2551,11 @@ class SpatialDataSetComparison:
         self.fractions = fractions
 
         if "legacy" in json_dict[self.ref_exp]["Analysis parameters"].keys():
-            self.markerproteins = json_dict[self.ref_exp]["complexes"]
+            self.markerproteins = json_dict[self.ref_exp][SettingStrings.COMPLEXES]
         else:
             try:
                 organism = json_dict[list(json_dict.keys())[0]]["Analysis parameters"][
-                    "organism"
+                    SettingStrings.ORGANISM
                 ]
             except:
                 organism = "Homo sapiens - Uniprot"
